fix: rename dashboard to projects
